requirements.txt
django==1.9.4
djangorestframework==3.3.3

-e ./algorithms

-r ./algorithms/requirements.txt